add: authentication token, fix: login page

+++ src/components/Adminpage/Hero.js
@@ -29,7 +29,17 @@ const Hero = () => {
     useEffect(() => {
         const fetchUsers = async () => {
             try {
-                const response = await axios.get(`${process.env.REACT_APP_API_URL}/users`);
+                const token = localStorage.getItem('token');
+                if (!token) {
+                    setError('Please login to access this page');
+                    setLoading(false);
+                }
+                const response = await axios.get(`${process.env.REACT_APP_API_URL}/users`, {
+                    headers: {
+                        'Content-Type': 'application/json',
+                        'Authorization': `Bearer ${token}`
+                    }
+                });
                 const filteredUsers = response.data?.filter(user => !user.role);
                 setUsers(filteredUsers);
                 setLoading(false);
@@ -59,10 +69,16 @@ const Hero = () => {
 
     const handleConfirmDelete = async () => {
         try {
+            const token = localStorage.getItem('token');
+            if (!token) {
+                setError('Please login to access this page');
+                return;
+            }
             await axios.delete(`${process.env.REACT_APP_API_URL}/deleteUser`, {
                 data: { userId: userIdToDelete },
                 headers: {
                     'Content-Type': 'application/json',
+                    'Authorization': `Bearer ${token}`
                 }
             });
             setUsers((prevUsers) => prevUsers?.filter(user => user.userId !== userIdToDelete));
@@ -79,7 +95,19 @@ const Hero = () => {
 
     const fetchFormDetails = async (userId) => {
         try {
-            const response = await axios.post(`${process.env.REACT_APP_API_URL}/search-forms`, { employeeId: userId });
+            const token = localStorage.getItem('token');
+            if (!token) {
+                setError('Please login to access this page');
+                return;
+            }
+            const response = await axios.post(`${process.env.REACT_APP_API_URL}/search-forms`, { employeeId: userId }, 
+                {
+                    headers: {
+                        'Content-Type': 'application/json',
+                        'Authorization': `Bearer ${token}`
+                    }
+                }
+            );
             setFormsFetched(response.data.forms);
         } catch (err) {
             setError(err.response.data.message);
